pyproject.toml
[project]
name = "utilitybill-pipeline"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "google-cloud-documentai",
  "google-cloud-storage",
  "google-cloud-bigquery",
  "google-cloud-pubsub",
  "google-generativeai",
  "pydantic>=2",
  "fastapi",
  "uvicorn[standard]",
  "celery[redis]",
  "python-multipart",
  "tqdm",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203","W503"]
